Separate extraction and parsing.
